requirements.txt
python-telegram-bot>=20,<21
aiohttp>=3.9
aiosqlite>=0.19
APScheduler>=3.10
python-dotenv>=1.0

